requirements.txt
certifi==2018.4.16
chardet==3.0.4
cloudinary==1.11.0
dj-database-url==0.5.0
Django==2.0.4
django-crispy-forms==1.7.2
django-environ==0.4.4
django-filebrowser-no-grappelli==3.7.4
django-js-asset==1.1.0
django-tinymce4-lite==1.7.1
gunicorn==19.7.1
idna==2.6
jsmin==2.2.2
mock==2.0.0
pbr==4.0.2
Pillow==5.1.0
pipenv==11.10.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycryptodomex==3.4.7
PyJWT==1.3.0
python-decouple==3.1
pytz==2018.4
requests==2.18.4
six==1.11.0
stream-django==1.4.0
stream-python==2.8.1
toml==0.9.4
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
